footer: add disclaimer about this blog
Everything in this blog is my craps, written in my most drunken state.
Everything in this blog is neither fact, nor advice.
Please take it with a grain of salt.
Don't trust me on anything.

THE BLOG IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE BLOG OR THE USE OR OTHER DEALINGS IN
THE BLOG.

Of course, NONE of my employers have anything to do with this blog.